+++ src/main/java/com/ll/hfback/domain/festival/repository/KopisRepository.java
@@ -1,6 +1,6 @@
-package li.com.backend_apiconnet.repository;
+package com.ll.hfback.domain.festival.repository;
 
-import li.com.backend_apiconnet.entity.KopisFesEntity;
+import com.ll.hfback.domain.festival.entity.KopisFesEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
