--- src/test/java/li/com/backend_apiconnet/BackendApiconnetApplicationTests.java
@@ -1,13 +1,0 @@
-package li.com.backend_apiconnet;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackendApiconnetApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
